- Fix drifting location when adding new POI to guardian sites - Stop PlotPriorScans showing when Guardian plotters are active - Use decimal not float in many places - Show Obelisk details in PlotGuardianStatus

+++ SrvSurvey/FormRuins.cs
@@ -475,7 +475,7 @@ namespace SrvSurvey
             {
                 // calculate render point for POI
                 var pt = Util.rotateLine(
-                    360 - poi.angle,
+                    360 - (decimal)poi.angle,
                     poi.dist);
 
                 // is this the closest POI?
